backend/requirements.txt
# Web framework & ASGI
fastapi==0.100.0
uvicorn[standard]==0.23.1
gunicorn==20.1.0

# Database / ORM
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
# Auth & config
PyJWT==2.8.0
python-dotenv==1.0.0

# AWS / S3
boto3>=1.26.0
botocore==1.38.9

# Background tasks & queue
celery==5.3.1
redis==4.6.0

# Uploads / forms
python-multipart==0.0.6
aiofiles==23.1.0
requests==2.31.0

# Ensure a consistent wheel environment
setuptools>=59.6.0
wheel
openpyxl==3.1.5
